Invoke scaffolding code directly
This way, we have more control over it. E.g. we could now pass
extra optional arguments to CreateTitlePage or CreateMainPage.

Moreover, we are now one step closer to using GenerateDocumentation
without using any of the primary AutoDoc features (the scaffolding
code is mostly independent from the rest of the AutoDoc code).